+++ runtimes/common/src/lib.rs
@@ -55,7 +55,7 @@ pub mod pallet_id;
 pub mod xcm_config;
 
 #[cfg(feature = "runtime-benchmarks")]
-pub mod benchmark;
+pub mod benchmarks;
 
 /// Opaque types. These are used by the CLI to instantiate machinery that don't
 /// need to know the specifics of the runtime. They can then be made to be
